tree: move git_oid into tree entry
A tree entry previously pointed directly into the object id within the
tree object itself; this is useful to avoid any unnecessary memory copy
(and an unnecessary use of 40 bytes per tree entry) but difficult if we
change the underlying `git_oid` object to not simply be a raw object id
but have additional structure.

This commit moves the `git_oid` directly into the tree entry; this
simplifies the tree entry creation from user data.  We now copy the
`git_oid` into place when parsing.